refactor(conventions): rewrite to independant plugin projects
