+++ gradle/conventions/global-conventions/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+    `kotlin-dsl`
+    `java-gradle-plugin`
+}
+
+dependencies {
+    implementation("com.github.ben-manes:gradle-versions-plugin:0.39.0")
+    implementation("net.linguica.gradle:maven-settings-plugin:0.5")
+}
+
+gradlePlugin {
+    plugins {
+        register("globalConventions") {
+            id = "$group.global-conventions"
+            displayName = "OpenAPI SDK conventions plugin"
+            description = "A plugin to common conventions"
+            implementationClass = "com.equisoft.standards.gradle.global.GlobalConventionsPlugin"
+        }
+    }
+}
