+++ gradle/conventions/openapi-sdk/src/main/kotlin/com/equisoft/standards/gradle/openapisdk/OpenApiSdkExtension.kt
@@ -1,0 +1,28 @@
+package com.equisoft.standards.gradle.openapisdk
+
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.file.ProjectLayout
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
+
+abstract class OpenApiSdkExtension(layout: ProjectLayout) {
+    abstract val customResourcesDir: DirectoryProperty
+    abstract val openApiPropertiesFile: RegularFileProperty
+    abstract val outputDir: DirectoryProperty
+    abstract val specFile: RegularFileProperty
+    abstract val swaggerVersion: Property<String?>
+
+    init {
+        outputDir
+            .convention(layout.buildDirectory.dir("sdk"))
+            .finalizeValueOnRead()
+        customResourcesDir
+            .convention(layout.projectDirectory.dir("build-resources/openapi-generator"))
+            .finalizeValueOnRead()
+        openApiPropertiesFile
+            .convention(layout.buildDirectory.file("openapi.properties"))
+            .finalizeValueOnRead()
+        specFile.finalizeValueOnRead()
+        swaggerVersion.finalizeValueOnRead()
+    }
+}
